Locates the currently running Nginx directory
In the CentOS, Nginx configuration is not located in ```/etc/nginx```. A more [direct solution](https://stackoverflow.com/questions/19910042/locate-the-nginx-conf-file-my-nginx-is-actually-using) is provided. Running the test command twice is required for the purpose of the exit status. For example: 

```
>/usr/bin/foo bar
-bash: /usr/bin/foo: No such file or directory
>echo $?
>/usr/bin/foo bar 2>&1
-bash: /usr/bin/foo: No such file or directory
>echo $?
0
>/usr/bin/foo && /usr/bin/foo 2>&1 | head -n1 | cut -d' ' -f5
-bash: /usr/bin/foo: No such file or directory
>echo $?
127
>test=$(/usr/nginx/sbin/nginx -t && /usr/nginx/sbin/nginx -t 2>&1 | head -n1 | cut -d' ' -f5)
nginx: the configuration file /usr/nginx/conf/nginx.conf syntax is ok
nginx: configuration file /usr/nginx/conf/nginx.conf test is successful
>echo $?
0
>echo $test
/usr/nginx/conf/nginx.conf
```