Make the user choice of master's job execution host status sticky
By default, when a worker instance is added, the master node will be
disabled from running jobs. This can be overridden (on the Admin page)
by toggling master's exec host status. Before this commit, the status
would be reset with each worker node being added. With this change,
the user choice will persist.

In part, this addresses issue #63.