Modify auto-scaling params to accommodate multi-core jobs
When a multi-core job is submitted to Slurm, and there are no
resources to fulfill it, the job would just sit in the queue with
a flag ReqNodeNotAvail. CloudMan's auto-scaling logic never picked
that up because the job was never actually queued. This modifies that
logic to accommodate such jobs. One user-facing requirement is to
choose instance type that has sufficient core capacity for such jobs.
If not, workers will be added but the jobs still won't be able to run
and in the mean time, auto-scaling will not continue to add any more
workers.

Partially addresses issue #60.